+++ src/interfaces/ICharacter.sol
@@ -8,7 +8,7 @@ interface ICharacter {
         uint8 health;
         uint8 strength;
         uint8 magic;
-        uint16 mana;
+        uint8 mana;
         address location_dungeon;
         string location_exact;
     }
